+++ internal/discovery/test_helpers_test.go
@@ -55,6 +55,7 @@ func (s stubExchange) GetFundingFees(string, time.Time) ([]exchange.FundingPayme
 	return nil, nil
 }
 func (s stubExchange) EnsureOneWayMode() error { return nil }
+func (s stubExchange) Close()                   {}
 func (s stubExchange) GetClosePnL(string, time.Time) ([]exchange.ClosePnL, error) {
 	return nil, nil
 }